start the users tables API

+++ backend/src/DB/db_service.py
@@ -6,17 +6,24 @@ import logging
 
 class DB:
     _client = None
-    _db_name = "smart_sheets"
+    _db_name = None
+    _collection_name: str = None
+    _coll_users = "users"
+    _coll_tables = "users_tables"
+    _coll_metadata = "users_metadata"
 
     @staticmethod
-    async def get_collection_db(db_name: str = "smart_sheets", collection_name: str = "users"):
+    async def get_collection_db(db_name: str = None, collection_name: str = None):
         if DB._client is None:
+            DB._db_name = db_name or DB._db_name or "smart_sheets"
+            DB._collection_name = collection_name or DB._collection_name or "users"
             try:
                 DB._client = AsyncIOMotorClient(Env.get_DB_port())
             except Exception as error:
                 logging.error(f"Error connecting to DB: {error}")
                 raise error
-        return DB._client[db_name][collection_name]
+
+        return DB._client[DB._db_name][DB._collection_name]
 
     @staticmethod
     async def close_db():
@@ -33,7 +40,7 @@ class DB:
     async def find(
         conditions: Optional[dict] = None,
         fields: Optional[dict] = None,
-        find_one: bool = False,
+        many: bool = False,
         sort_fields: Optional[List[Tuple[str, int]]] = None,
         limit: Optional[int] = None,
         skip: Optional[int] = None,
@@ -42,15 +49,15 @@ class DB:
         collection_name: str = None,
     ) -> Union[dict, list]:
 
-        db_name = db_name or DB._db_name
-        collection_name = collection_name or "users"
+        DB._db_name = db_name or DB._db_name
+        DB._collection_name = collection_name or DB._collection_name
+        db = await DB.get_collection_db()
+
         conditions = conditions or {}
         fields = fields or {}
         additional_options = additional_options or {}
 
-        db = await DB.get_collection_db(db_name=db_name, collection_name=collection_name)
-
-        if find_one:
+        if not many:
             return await db.find_one(conditions, fields)
 
         cursor = db.find(conditions, fields)
@@ -71,25 +78,27 @@ class DB:
 
     @staticmethod
     async def insert(
-        documents: List[dict] = None,
-        insert_one: bool = False,
-        write_concern: Optional[dict] = None,
+        # תמיכה במסמך יחיד או ברשימה
+        documents: Union[dict, List[dict]] = None,
+        many: bool = False,
         db_name: str = None,
         collection_name: str = None,
     ) -> Union[dict, list]:
 
-        db_name = db_name or DB._db_name
-        collection_name = collection_name or "users"
+        DB._db_name = db_name or DB._db_name
+        DB._collection_name = collection_name or DB._collection_name
+        db = await DB.get_collection_db()
 
         if documents is None:
-            documents = [{}] if insert_one else []
+            documents = [{}] if not many else []
 
-        db = await DB.get_collection_db(db_name=db_name, collection_name=collection_name)
+        if not isinstance(documents, list):
+            documents = [documents]
 
         try:
-            if insert_one:
+            if not many:
                 result = await db.insert_one(documents[0])
-                return {**documents[0], "_id": str(result.inserted_id)}
+                return {"_id": str(result.inserted_id), **documents[0]}
 
             results = await db.insert_many(documents)
             return [{"_id": str(inserted_id)} for inserted_id in results.inserted_ids]
@@ -102,34 +111,40 @@ class DB:
         conditions: Optional[dict] = None,
         updates: Optional[dict] = None,
         replace: Optional[dict] = None,
-        update_one: bool = False,
+        array_updates: Optional[dict] = None,
+        array_removals: Optional[dict] = None,
+        increase: Optional[dict] = None,
         upsert: bool = False,
+        many: bool = False,
         db_name: str = None,
         collection_name: str = None,
     ) -> Union[dict, list]:
 
-        db_name = db_name or "smart_sheets"
-        collection_name = collection_name or "users"
-
-        if conditions is None:
-            conditions = {}
-        if updates is None:
-            updates = {}
-
-        if replace is None:
-            replace = {}
-
-        db = await DB.get_collection_db(db_name=db_name, collection_name=collection_name)
+        DB._db_name = db_name or DB._db_name
+        DB._collection_name = collection_name or DB._collection_name
+        db = await DB.get_collection_db()
 
         try:
-            if update_one:
-                result = await db.update_one(conditions, {"$set": updates, "$unset": replace}, upsert=upsert)
-                if result.modified_count > 0:
-                    return {"message": "Document updated successfully", "status": True}
-                return {"message": "No documents matched the query", "status": False}
+            update_operations = {
+                "$set": updates or {}, "$unset": replace or {}}
 
-            result = await db.update_many(conditions, {"$set": updates, "$unset": replace}, upsert=upsert)
-            return {"modified_count": result.modified_count, "status": True}
+            if array_updates:
+                update_operations["$push"] = array_updates
+
+            if array_removals:
+                update_operations["$pull"] = array_removals
+
+            if increase:
+                update_operations["$inc"] = increase
+
+            if not many:
+                result = await db.update_one(conditions or {}, update_operations, upsert=upsert)
+            else:
+                result = await db.update_many(conditions or {}, update_operations, upsert=upsert)
+
+            if result.modified_count > 0:
+                return {"message": "Document updated successfully", "status": True}
+            return {"message": "No documents matched the query", "status": False}
         except Exception as e:
             logging.error(f"Update failed: {e}")
             raise
@@ -137,28 +152,28 @@ class DB:
     @staticmethod
     async def delete(
         conditions: Optional[dict] = None,
-        delete_one: bool = False,
+        many: bool = False,
         db_name: str = None,
         collection_name: str = None,
     ) -> dict:
 
-        db_name = db_name or "smart_sheets"
-        collection_name = collection_name or "users"
+        DB._db_name = db_name or DB._db_name
+        DB._collection_name = collection_name or DB._collection_name
+        db = await DB.get_collection_db()
 
         if conditions is None:
             conditions = {}
 
-        db = await DB.get_collection_db(db_name=db_name, collection_name=collection_name)
-
         try:
-            if delete_one:
+            if not many:
                 result = await db.delete_one(conditions)
-                if result.deleted_count > 0:
-                    return {"message": "Document deleted successfully"}
-                return {"message": "No document matched the query"}
+            else:
+                result = await db.delete_many(conditions)
 
-            result = await db.delete_many(conditions)
-            return {"message": f"{result.deleted_count} documents deleted successfully"}
+            if result.deleted_count > 0:
+                return {"message": f"{result.deleted_count} documents deleted successfully"}
+            return {"message": "No document matched the query"}
+
         except Exception as e:
             logging.error(f"Delete failed: {e}")
             raise
@@ -172,7 +187,9 @@ class DB:
         if not isinstance(index_fields, list) or not all(isinstance(field, tuple) for field in index_fields):
             raise ValueError("index_fields must be a list of tuples")
 
-        db = await DB.get_collection_db(db_name=db_name, collection_name=collection_name)
+        DB._db_name = db_name or DB._db_name
+        DB._collection_name = collection_name or DB._collection_name
+        db = await DB.get_collection_db()
 
         try:
             result = await db.create_index(index_fields)
